Improved send post
