+++ src/components/Avatar/Avatar.tsx
@@ -1,16 +1,19 @@
 import { TAvatarArray } from '../../../src/hooks/useAvatarArray';
-import useAuth from '../../../src/hooks/useAuth';
 import { AVATAR_SIZE, useAvatarStyles } from './Avatar.styles';
 import * as React from 'react';
 import { View, Text, Image, FlatList } from 'react-native';
 
 
 export type TAvatar = {
-    avatars: TAvatarArray[]
+    avatars: TAvatarArray[];
+    heightProp?: number;
+    widthProp?: number;
 }
 
-export const Avatar = ({ avatars }: TAvatar) => {
-    const { apiRegion } = useAuth();
+const DEFAULT_WIDTH = AVATAR_SIZE;
+const DEFAULT_HEIGHT = AVATAR_SIZE;
+
+export const Avatar = ({ avatars, heightProp = DEFAULT_HEIGHT, widthProp = DEFAULT_WIDTH }: TAvatar) => {
     const styles = useAvatarStyles();
     const renderAvatar = (avatar: TAvatarArray, index: number, height: number, width: number) => {
         switch (avatar.type) {
@@ -21,7 +24,8 @@ export const Avatar = ({ avatars }: TAvatar) => {
                         height: height,
                     }]} key={index}>
                         <View style={{ width: '100%', height: '100%' }}>
-                            <Image resizeMode='cover' source={{ uri: `https://api.${apiRegion}.amity.co/api/v3/files/${avatar.value}/download?size=small` }} style={styles.avatarImage} />
+                            {/* TODO: Replace harcoded "us" with apiregion  */}
+                            <Image resizeMode='cover' source={{ uri: `https://api.${'us'}.amity.co/api/v3/files/${avatar.value}/download?size=small` }} style={styles.avatarImage} />
                         </View>
                     </View>
                 );
@@ -59,14 +63,14 @@ export const Avatar = ({ avatars }: TAvatar) => {
                         <View style={{ width: '50%', borderRightWidth: 0.5, borderColor: '#fff', }}>
                             {
                                 // @ts-ignore
-                                renderAvatar(avatars[0], 0, AVATAR_SIZE, AVATAR_SIZE / 2)
+                                renderAvatar(avatars[0], 0, heightProp, widthProp / 2)
                             }
                         </View>
                         <View style={{ width: '50%', overflow: 'hidden' }}>
                             <FlatList
                                 data={avatars.slice(1, avatars.length)}
                                 numColumns={1}
-                                renderItem={({ item, index }) => renderAvatar(item, index, AVATAR_SIZE / 2, AVATAR_SIZE / 2)}
+                                renderItem={({ item, index }) => renderAvatar(item, index, heightProp / 2, widthProp / 2)}
                                 keyExtractor={(item) => item.value.toString()}
                             />
                         </View>
@@ -74,7 +78,7 @@ export const Avatar = ({ avatars }: TAvatar) => {
                 ) : <FlatList
                     data={avatars}
                     numColumns={2}
-                    renderItem={({ item, index }) => renderAvatar(item, index, AVATAR_SIZE / 2, AVATAR_SIZE / 2)}
+                    renderItem={({ item, index }) => renderAvatar(item, index, heightProp / 2, widthProp / 2)}
                     keyExtractor={(item) => item.value.toString()}
                 />
             }
