+++ src/screens/ChatRoom/ChatRoom.tsx
@@ -7,6 +7,7 @@ import React, {
   useRef,
   useLayoutEffect,
   useContext,
+  useCallback,
 } from 'react';
 import {
   View,
@@ -15,11 +16,9 @@ import {
   TouchableOpacity,
   TextInput,
   Platform,
-  Text,
   KeyboardAvoidingView,
   FlatList,
   Keyboard,
-  Alert,
   ActivityIndicator,
   LayoutAnimation,
 } from 'react-native';
@@ -36,7 +35,6 @@ import {
 import type { RootStackParamList } from '../../routes/RouteParamList';
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import BackButton from '../../components/BackButton';
-import moment from 'moment';
 import {
   MessageContentType,
   MessageRepository,
@@ -54,14 +52,6 @@ import ImagePicker, {
 } from 'react-native-image-picker';
 // import { SafeAreaView } from 'react-native-safe-area-context';
 import LoadingImage from '../../components/LoadingImage';
-import {
-  Menu,
-  MenuOptions,
-  MenuOption,
-  MenuTrigger,
-} from 'react-native-popup-menu';
-import { SvgXml } from 'react-native-svg';
-import { deletedIcon } from '../../svg/svg-xml-list';
 import EditMessageModal from '../../components/EditMessageModal';
 // import { GroupChatIcon } from '../../svg/GroupChatIcon';
 import { AvatarIcon } from '../../svg/AvatarIcon';
@@ -78,6 +68,7 @@ import { AuthContext } from '../../store/context';
 import { useReadStatus } from '../../hooks/useReadStatus';
 import { useAvatarArray } from '../../../src/hooks/useAvatarArray';
 import { Avatar } from '../../../src/components/Avatar/Avatar';
+import { EachChatMessage } from '@amityco/react-native-cli-chat-ui-kit/src/screens/ChatRoom/EachChatMessage';
 
 type ChatRoomScreenComponentType = React.FC<{}>;
 LogBox.ignoreLogs(['Warning: ...']); // Ignore log notification by message
@@ -87,7 +78,17 @@ export enum ECustomData {
   announcement = 'announcement',
   post = 'post'
 }
-interface IMessage {
+
+export type TCustomData = {
+  type?: ECustomData
+  text?: string;
+  imageIds?: string[];
+  id?: string;
+  extraData: {
+    postCreator: Amity.User
+  }
+}
+export interface IMessage {
   _id: string;
   text?: string;
   createdAt: string;
@@ -101,12 +102,7 @@ interface IMessage {
   messageType: string;
   isPending?: boolean;
   isDeleted: boolean;
-  customData?: {
-    type?: ECustomData
-    text?: string;
-    images?: never[];
-    id?: string;
-  };
+  customData?: TCustomData;
 }
 export interface IDisplayImage {
   url: string;
@@ -276,12 +272,16 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
             },
             messageType: item.dataType,
             isDeleted: item.isDeleted as boolean,
+            // @ts-ignore
             customData: {
               type: '',
               text: '',
-              images: [],
-              id: ''
-            }
+              imageIds: [],
+              id: '',
+              extraData: {
+                postCreator: null
+              }
+            } as TCustomData
           };
           // @ts-ignore
           if (item.dataType === 'custom' && item?.data?.type === ECustomData.post && item?.data?.id) {
@@ -293,15 +293,10 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
               if (data) {
                 //let imageUrls = []
                 if (data.children?.length > 0) {
-                  //if post has any images children, fetch image data from childrenPostId
-                  for (const item of data.children) {
-                    PostRepository.getPost(item.data?.postId, ({ data }) => {
-                      console.log("data for children post", data)
-                      if (data) {
-
-                      }
-                    });
-                  }
+                  commonObj.customData.imageIds = [...data.children]
+                }
+                if (data?.creator && Object.keys(data.creator).length > 0) {
+                  commonObj.customData.extraData.postCreator = data?.creator
                 }
                 commonObj.customData.type = ECustomData.post
                 commonObj.customData.id = data._id;
@@ -392,293 +387,6 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
     const reOrderArr = sortedMessagesData;
     setSortedMessages([...reOrderArr]);
   }, [messages]);
-
-  const openFullImage = (image: string, messageType: string) => {
-    if (messageType === 'image' || messageType === 'file') {
-      const fullSizeImage: string = image + '?size=full';
-      setFullImage(fullSizeImage);
-      setIsVisibleFullImage(true);
-    }
-  };
-
-  const renderTimeDivider = (date: string) => {
-    const currentDate = date;
-    const formattedDate = moment(currentDate).format('MMMM DD, YYYY');
-    const today = moment().startOf('day');
-
-    let displayText = formattedDate;
-
-    if (moment(currentDate).isSame(today, 'day')) {
-      displayText = 'Today';
-    }
-
-    return (
-      <View style={styles.bubbleDivider}>
-        <View style={styles.textDivider}>
-          <Text style={styles.dateText}>{displayText}</Text>
-        </View>
-      </View>
-    );
-  };
-
-  const deleteMessage = async (messageId: string) => {
-    const message = await MessageRepository.softDeleteMessage(messageId);
-    return message;
-  };
-
-  const reportMessage = async (messageId: string) => {
-    const isFlagged = await MessageRepository.flagMessage(messageId);
-    if (isFlagged) {
-      Alert.alert('Report sent ✅');
-    }
-  };
-
-  const renderChatMessages = (message: IMessage, index: number) => {
-    const isUserChat: boolean =
-      message?.user?._id === (client as Amity.Client).userId;
-    //isUserChat - is chat of the the user who is logged in?
-
-    let isRenderDivider = false;
-    const messageDate = moment(message.createdAt);
-
-    const previousMessageDate = moment(sortedMessages[index + 1]?.createdAt);
-    const isSameDay = messageDate.isSame(previousMessageDate, 'day');
-
-    if (!isSameDay || index === sortedMessages.length - 1) {
-      isRenderDivider = true;
-    }
-
-    const isNormalText = message.messageType === 'text';
-    const isImage = message.messageType === 'image';
-    const isAnnouncement = message?.messageType === "custom" && message.customData?.type === ECustomData.announcement
-    const isSocialPost = message?.messageType === "custom" && message.customData?.type === ECustomData.post
-
-    return (
-      <View>
-        {isRenderDivider && renderTimeDivider(message.createdAt)}
-        <View
-          style={!isUserChat ? isAnnouncement ? styles.leftMessageWrap : [styles.leftMessageWrap, { flexDirection: 'row' }] : styles.rightMessageWrap}
-        >
-          {!isUserChat && !isAnnouncement &&
-            (message.user.avatar ? (
-              <Image
-                source={{ uri: message.user.avatar }}
-                style={styles.avatarImage}
-              />
-            ) : (
-              <View style={styles.avatarImage}>
-                <AvatarIcon />
-              </View>
-            ))}
-
-          <View>
-            {!isUserChat && isGroupChat && !isAnnouncement ? (
-              <Text
-                style={isUserChat ? styles.chatUserText : styles.chatFriendText}
-              >
-                {message.user.name}
-              </Text>
-            ) : null}
-            {isAnnouncement && message.customData?.text ? (
-              <Text style={{ color: '#6E768A', fontSize: 12, fontWeight: '400', marginHorizontal: 20, lineHeight: 17, alignSelf: 'center' }}>
-                {message.customData?.text}
-              </Text>
-            ) : null}
-            {message.isDeleted ? (
-              <View
-                style={[
-                  styles.deletedMessageContainer,
-                  isUserChat
-                    ? styles.userMessageDelete
-                    : styles.friendMessageDelete,
-                ]}
-              >
-                <View style={styles.deletedMessageRow}>
-                  <SvgXml xml={deletedIcon} width={20} height={20} />
-                  <Text style={styles.deletedMessage}>Message Deleted</Text>
-                </View>
-              </View>
-            ) : (
-              <Menu>
-                <MenuTrigger
-                  onAlternativeAction={() =>
-                    openFullImage(message.image as string, message.messageType)
-                  }
-                  customStyles={{
-                    triggerTouchable: { underlayColor: 'transparent' },
-                  }}
-                  triggerOnLongPress
-                >
-                  {isNormalText ? (
-                    <View
-                      key={message._id}
-                      style={[
-                        styles.textChatBubble,
-                        isUserChat ? styles.userBubble : styles.friendBubble,
-                        isGroupChat
-                          ? { marginVertical: 5 }
-                          : { marginBottom: 5 },
-                      ]}
-                    >
-                      <Text
-                        style={
-                          isUserChat
-                            ? styles.chatUserText
-                            : styles.chatFriendText
-                        }
-                      >
-                        {message.text}
-                      </Text>
-                    </View>
-                  ) : isImage ? (
-                    <View
-                      style={[
-                        styles.imageChatBubble,
-                        isUserChat
-                          ? styles.userImageBubble
-                          : styles.friendBubble,
-                      ]}
-                    >
-                      <Image
-                        style={styles.imageMessage}
-                        source={{
-                          uri: message.image + '?size=medium',
-                        }}
-                      />
-                    </View>
-                  ) : isSocialPost ? (
-                    <>
-                      {
-                        message.customData?.images && message.customData.images?.length > 0 ? (
-                          message.customData.images?.map((eachImageUrl) => {
-                            return (
-                              <View
-                                style={[
-                                  styles.imageChatBubble,
-                                  isUserChat
-                                    ? styles.userImageBubble
-                                    : styles.friendBubble,
-                                ]}
-                              >
-                                <Image
-                                  style={styles.imageMessage}
-                                  source={{
-                                    uri: eachImageUrl + '?size=medium',
-                                  }}
-                                />
-                              </View>
-                            )
-                          })
-                        ) : null
-                      }
-                      {
-                        message.customData?.text ? (
-                          <View
-                            key={message._id}
-                            style={[
-                              styles.textChatBubble,
-                              isUserChat ? styles.userBubble : styles.friendBubble,
-                              isGroupChat
-                                ? { marginVertical: 5 }
-                                : { marginBottom: 5 },
-                            ]}
-                          >
-                            <Text
-                              style={
-                                isUserChat
-                                  ? styles.chatUserText
-                                  : styles.chatFriendText
-                              }
-                            >
-                              {message.customData.text}
-                            </Text>
-                          </View>
-                        ) : null
-                      }
-                    </>
-                  ) : null}
-                </MenuTrigger>
-                <MenuOptions
-                  customStyles={{
-                    optionsContainer: {
-                      ...styles.optionsContainer,
-                      marginLeft: isUserChat
-                        ? 240 +
-                        (message.text && message.text.length < 5
-                          ? message.text.length * 10
-                          : 10)
-                        : 0,
-                    },
-                  }}
-                >
-                  {isUserChat ? (
-                    <MenuOption
-                      onSelect={() =>
-                        Alert.alert(
-                          'Delete this message?',
-                          `Message will be also be permanently removed from your friend's devices.`,
-                          [
-                            { text: 'Cancel', style: 'cancel' },
-                            {
-                              text: 'Delete',
-                              style: 'destructive',
-                              onPress: () => deleteMessage(message._id),
-                            },
-                          ]
-                        )
-                      }
-                      text="Delete"
-                    />
-                  ) : (
-                    <MenuOption
-                      onSelect={() => reportMessage(message._id)}
-                      text="Report"
-                    />
-                  )}
-                  {message.messageType === 'text' && isUserChat && (
-                    <MenuOption
-                      onSelect={() => {
-                        return openEditMessageModal(
-                          message._id,
-                          message.text as string
-                        );
-                      }}
-                      text="Edit"
-                    />
-                  )}
-                </MenuOptions>
-              </Menu>
-            )}
-            {
-              !isAnnouncement ? (
-                <View
-                  style={{
-                    flexDirection: 'row',
-                    alignSelf: isUserChat ? 'flex-end' : 'flex-start',
-                  }}
-                >
-                  <Text
-                    style={[
-                      styles.chatTimestamp,
-                      {
-                        alignSelf: isUserChat ? 'flex-end' : 'flex-start',
-                      },
-                    ]}
-                  >
-                    {message.createdAt != message.editedAt ? 'Edited ·' : ''}{' '}
-                    {moment(message.createdAt).format('hh:mm A')}
-                  </Text>
-                  {isUserChat && isDelivered ?
-                    getReadComponent(message._id)
-                    : null}
-                </View>
-              ) : null
-            }
-          </View>
-        </View>
-      </View>
-    );
-  };
 
   const handlePress = () => {
     Keyboard.dismiss();
@@ -797,17 +505,25 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
     );
   }, [displayImages, handleOnFinishImage]);
 
-  const openEditMessageModal = (messageId: string, text: string) => {
+  const openEditMessageModal = useCallback((messageId: string, text: string) => {
     setEditMessageId(messageId);
     setEditMessageModal(true);
     setEditMessageText(text);
-  };
+  }, []);
 
   const closeEditMessageModal = () => {
     setEditMessageId('');
     setEditMessageText('');
     setEditMessageModal(false);
   };
+
+  const openFullImage = useCallback((image: string, messageType: string) => {
+    if (messageType === 'image' || messageType === 'file') {
+      const fullSizeImage: string = image + '?size=full';
+      setFullImage(fullSizeImage);
+      setIsVisibleFullImage(true);
+    }
+  }, []);
 
   return (
     <View style={styles.container}>
@@ -868,7 +584,16 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
       <View style={styles.chatContainer}>
         <FlatList
           data={sortedMessages}
-          renderItem={({ item, index }) => renderChatMessages(item, index)}
+          renderItem={({ item, index }) => (
+            <EachChatMessage
+              isGroupChat={isGroupChat}
+              index={index}
+              message={item}
+              sortedMessages={sortedMessages}
+              openEditMessageModal={openEditMessageModal}
+              openFullImage={openFullImage}
+            />
+          )}
           keyExtractor={(item) => item._id}
           onEndReached={loadNextMessages}
           onEndReachedThreshold={0.5}
